+++ addon/services/form-load-time-tracker.js
@@ -1,13 +1,71 @@
+/**
+  @module ember-flexberry
+*/
+
 import Ember from 'ember';
 
+/**
+  Service for store time load and render.
+
+  @class FormLoadTimeTrackerService
+  @extends Ember.Service
+*/
 export default Ember.Service.extend({
-  initTime: 0,
-  beforeModelTime: 0,
-  modelTime: 0,
-  afterModelTime: 0,
-  activateTime: 0,
-  setupControllerTime: 0,
-  renderTemplateTime:0,
-  firstDidRenderTime: 0,
-  lastDidRenderTime: 0
+  /**
+    Start time of load data.
+
+    @property startLoadTime
+    @type Number
+    @default 0
+  */
+  startLoadTime: 0,
+
+  /**
+    End time of load data.
+
+    @property endLoadTime
+    @type Number
+    @default 0
+  */
+  endLoadTime: 0,
+
+  /**
+    Time of start render.
+
+    @property startRenderTime
+    @type Number
+    @default 0
+  */
+  startRenderTime: 0,
+
+  /**
+    Time of end render.
+
+    @property endRenderTime
+    @type Number
+    @default 0
+  */
+  endRenderTime: 0,
+
+  /**
+    Time of load data.
+
+    @property loadTime
+    @type Number
+    @readonly
+  */
+  loadTime: Ember.computed('startLoadTime', 'endLoadTime', function() {
+    return this.get('endLoadTime') - this.get('startLoadTime');
+  }).readOnly(),
+
+  /**
+    Time of render.
+
+    @property renderTime
+    @type Number
+    @readonly
+  */
+  renderTime: Ember.computed('startRenderTime', 'endRenderTime', function() {
+    return this.get('endRenderTime') - this.get('startRenderTime');
+  }).readOnly(),
 });